requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
jinja2==3.1.2
python-multipart==0.0.6
supabase==2.0.3
pydantic==2.4.2
python-dotenv==1.0.0
reportlab==4.0.7
starlette==0.27.0
typing-extensions==4.8.0
python-jose==3.3.0
passlib==1.7.4
requests==2.31.0